refactor: move SQLite cost model tests to zero-cache to resolve cyclic dependency
